AI
LaunchBall destroys enemies when colliding.
Enemy AI needs tag "Enemy" for it to work.

Player takes damage when it collides with Enemy.
Player needs tag "Player" for it to work.

AI will start to walk to the right until it leaves a trigger box. When it leaves a trigger box it turns left until it leaves another trigger box.